make gcp id's unique per day fetching
It turns out that the StartTime in the GCP csv file doesn't have to be from the same month as when the file is reporting billing for. This can cause unexpected duplicate id's found. This way there will be a unique id per day of fetching data so it should always be unique.